Fix tab and space not working in WYSIWYG
Disclaimer: I translated this code back and forth from typescript.

I have tested this on Legacy mode and it works, but these edits fix it for latest version. The code for legacy is essentially copied from current.

Fixes... a whole bunch of issues. If I remember I'll edit and list them all, can't check right now.

The way this works is by using `this.registerDomEvent` instead of CodeMirror edits. The key down event now also checks whether the current window is a markdown view, so you can't have accidental replaces in any other areas. If that's actually wanted, feel free to remove it (or make it a setting, although I haven't seen a place where the distinction would actually be relevant)